refactor: auth/user 통합 및 디렉토리 정리
