+++ src/main/java/TiCatch/backend/domain/user/dto/response/LoginResponseDto.java
@@ -1,6 +1,6 @@
-package TiCatch.backend.domain.auth.dto.response;
+package TiCatch.backend.domain.user.dto.response;
 
-import TiCatch.backend.domain.auth.dto.TokenDto;
+import TiCatch.backend.domain.user.dto.TokenDto;
 import lombok.Builder;
 import lombok.Getter;
