+++ src/main/java/TiCatch/backend/global/util/HeaderUtil.java
@@ -1,6 +1,6 @@
-package TiCatch.backend.domain.auth.util;
+package TiCatch.backend.global.util;
 
-import TiCatch.backend.domain.auth.dto.TokenDto;
+import TiCatch.backend.domain.user.dto.TokenDto;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
